Value	Use a 64 bit variable (qint64) for integer values.
	Does not need more space as before, because double is always 64 bit and
	it is 'unioned' with double.

	This lifts the restriction to 32 bit for the bit operation functions.

CCMAIL: bradh@frogmouth.net

svn path=/trunk/koffice/; revision=628771

+++ kspread/Formula.h
@@ -168,7 +168,7 @@ class KSPREAD_EXPORT Token
      * Returns integer value for an integer token.
      * For any other type of token, returns 0.
      */
-    int asInteger() const;
+    qint64 asInteger() const;
 
     /**
      * Returns floating-point value for a floating-point token.
